Tighten up wait_us
The generic wait_ms, offered by the codal system timer, is not tight enough to perform small waits accurately. For instance, the sync call in the timer takes 4us alone to read the counter value and update the timestamp. This commit introduces an instruction based wait, using the clock speed of the CPU.